+++ src/components/Layout.tsx
@@ -1,5 +1,5 @@
 import React, { FC, ReactNode } from 'react'
-import styles from '~/styles/Layout.module.css'
+import styles from '~/styles/Layout.module.scss'
 import { staticPath } from '~/utils/$path'
 import DefaultHeader from './DefaultHeader'
