+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '~/styles/globals.css'
+import '~/styles/globals.scss'
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
 import { staticPath } from '~/utils/$path'
